// src/app/shared/models/student.ts
export interface Istudent{
    fname: string;
    lname: string;
    contact : string;
    email: string;
    id : string;
}

export interface Itodo{    
    name : string;
    id :string;       
}

export interface Icards{
    title: string;
    content: string;
    img: string;
}

export interface Inature{
    title: string;
    content: string;
    img: string;
}

export interface Ispace{
    title: string;
    content: string;
    img: string;
}